Update code to 4.5 official version

+++ python_moondb/python_moondb/api/policy.py
@@ -40,6 +40,9 @@ class PolicyManager(Managers):
     def update_policy(self, user_id, policy_id, value):
         if policy_id not in self.driver.get_policies(policy_id=policy_id):
             raise exceptions.PolicyUnknown
+        if value and 'model_id' in value and value['model_id'] != "":
+            if not Managers.ModelManager.get_models(user_id, model_id=value['model_id']):
+                raise exceptions.ModelUnknown
         return self.driver.update_policy(policy_id=policy_id, value=value)
 
     @enforce(("read", "write"), "policies")
@@ -60,6 +63,9 @@ class PolicyManager(Managers):
             raise exceptions.PolicyExisting
         if not policy_id:
             policy_id = uuid4().hex
+        if value and 'model_id' in value and value['model_id'] != "":
+            if not Managers.ModelManager.get_models(user_id, model_id=value['model_id']):
+                raise exceptions.ModelUnknown
         return self.driver.add_policy(policy_id=policy_id, value=value)
 
     @enforce("read", "policies")
@@ -68,10 +74,19 @@ class PolicyManager(Managers):
 
     @enforce("read", "perimeter")
     def get_subjects(self, user_id, policy_id, perimeter_id=None):
+        if not policy_id:
+            pass
+        elif not (policy_id and self.get_policies(user_id=user_id, policy_id=policy_id)):
+            raise exceptions.PolicyUnknown
         return self.driver.get_subjects(policy_id=policy_id, perimeter_id=perimeter_id)
 
     @enforce(("read", "write"), "perimeter")
     def add_subject(self, user_id, policy_id, perimeter_id=None, value=None):
+        if not value or "name" not in value or not value["name"].strip():
+            raise exceptions.PerimeterNameInvalid
+        if value["name"] in map(lambda x: x['name'],
+                                self.get_subjects(user_id, policy_id, perimeter_id).values()):
+            raise exceptions.SubjectExisting
         k_user = Managers.KeystoneManager.get_user_by_name(value.get('name'))
         if not k_user['users']:
             k_user = Managers.KeystoneManager.create_user(value)
@@ -94,10 +109,16 @@ class PolicyManager(Managers):
 
     @enforce(("read", "write"), "perimeter")
     def delete_subject(self, user_id, policy_id, perimeter_id):
+        if policy_id and not self.get_policies(user_id=user_id, policy_id=policy_id):
+            raise exceptions.PolicyUnknown
         return self.driver.delete_subject(policy_id=policy_id, perimeter_id=perimeter_id)
 
     @enforce("read", "perimeter")
     def get_objects(self, user_id, policy_id, perimeter_id=None):
+        if not policy_id:
+            pass
+        elif not (policy_id and self.get_policies(user_id=user_id, policy_id=policy_id)):
+            raise exceptions.PolicyUnknown
         return self.driver.get_objects(policy_id=policy_id, perimeter_id=perimeter_id)
 
     @enforce(("read", "write"), "perimeter")
@@ -110,21 +131,30 @@ class PolicyManager(Managers):
 
     @enforce(("read", "write"), "perimeter")
     def delete_object(self, user_id, policy_id, perimeter_id):
+        if policy_id and not self.get_policies(user_id=user_id, policy_id=policy_id):
+            raise exceptions.PolicyUnknown
         return self.driver.delete_object(policy_id=policy_id, perimeter_id=perimeter_id)
 
     @enforce("read", "perimeter")
     def get_actions(self, user_id, policy_id, perimeter_id=None):
+        if not policy_id:
+            pass
+        elif not (policy_id and self.get_policies(user_id=user_id, policy_id=policy_id)):
+            raise exceptions.PolicyUnknown
         return self.driver.get_actions(policy_id=policy_id, perimeter_id=perimeter_id)
 
     @enforce(("read", "write"), "perimeter")
     def add_action(self, user_id, policy_id, perimeter_id=None, value=None):
-        logger.info("add_action {}".format(policy_id))
+        logger.debug("add_action {}".format(policy_id))
         if not self.get_policies(user_id=user_id, policy_id=policy_id):
             raise exceptions.PolicyUnknown
         return self.driver.set_action(policy_id=policy_id, perimeter_id=perimeter_id, value=value)
 
     @enforce(("read", "write"), "perimeter")
     def delete_action(self, user_id, policy_id, perimeter_id):
+        logger.debug("delete_action {} {} {}".format(policy_id, perimeter_id, self.get_policies(user_id=user_id, policy_id=policy_id)))
+        if policy_id and not self.get_policies(user_id=user_id, policy_id=policy_id):
+            raise exceptions.PolicyUnknown
         return self.driver.delete_action(policy_id=policy_id, perimeter_id=perimeter_id)
 
     @enforce("read", "data")
@@ -144,6 +174,8 @@ class PolicyManager(Managers):
     def set_subject_data(self, user_id, policy_id, data_id=None, category_id=None, value=None):
         if not category_id:
             raise Exception('Invalid category id')
+        if not Managers.ModelManager.get_subject_categories(user_id=user_id, category_id=category_id):
+            raise exceptions.MetaDataUnknown
         if not self.get_policies(user_id=user_id, policy_id=policy_id):
             raise exceptions.PolicyUnknown
         if not data_id:
@@ -175,6 +207,8 @@ class PolicyManager(Managers):
     def add_object_data(self, user_id, policy_id, data_id=None, category_id=None, value=None):
         if not category_id:
             raise Exception('Invalid category id')
+        if not Managers.ModelManager.get_object_categories(user_id=user_id, category_id=category_id):
+            raise exceptions.MetaDataUnknown
         if not self.get_policies(user_id=user_id, policy_id=policy_id):
             raise exceptions.PolicyUnknown
         if not data_id:
@@ -206,6 +240,8 @@ class PolicyManager(Managers):
     def add_action_data(self, user_id, policy_id, data_id=None, category_id=None, value=None):
         if not category_id:
             raise Exception('Invalid category id')
+        if not Managers.ModelManager.get_action_categories(user_id=user_id, category_id=category_id):
+            raise exceptions.MetaDataUnknown
         if not self.get_policies(user_id=user_id, policy_id=policy_id):
             raise exceptions.PolicyUnknown
         if not data_id:
@@ -228,6 +264,12 @@ class PolicyManager(Managers):
     def add_subject_assignment(self, user_id, policy_id, subject_id, category_id, data_id):
         if not self.get_policies(user_id=user_id, policy_id=policy_id):
             raise exceptions.PolicyUnknown
+        if not self.get_subjects(user_id=user_id, policy_id=policy_id, perimeter_id=subject_id):
+            raise exceptions.SubjectUnknown
+        if not Managers.ModelManager.get_subject_categories(user_id=user_id, category_id=category_id):
+            raise exceptions.MetaDataUnknown
+        if not self.get_subject_data(user_id=user_id, policy_id=policy_id, data_id=data_id):
+            raise exceptions.DataUnknown
         return self.driver.add_subject_assignment(policy_id=policy_id, subject_id=subject_id,
                                                   category_id=category_id, data_id=data_id)
 
@@ -244,6 +286,12 @@ class PolicyManager(Managers):
     def add_object_assignment(self, user_id, policy_id, object_id, category_id, data_id):
         if not self.get_policies(user_id=user_id, policy_id=policy_id):
             raise exceptions.PolicyUnknown
+        if not self.get_objects(user_id=user_id, policy_id=policy_id, perimeter_id=object_id):
+            raise exceptions.ObjectUnknown
+        if not Managers.ModelManager.get_object_categories(user_id=user_id, category_id=category_id):
+            raise exceptions.MetaDataUnknown
+        if not self.get_object_data(user_id=user_id, policy_id=policy_id, data_id=data_id):
+            raise exceptions.DataUnknown
         return self.driver.add_object_assignment(policy_id=policy_id, object_id=object_id,
                                                  category_id=category_id, data_id=data_id)
 
@@ -260,6 +308,12 @@ class PolicyManager(Managers):
     def add_action_assignment(self, user_id, policy_id, action_id, category_id, data_id):
         if not self.get_policies(user_id=user_id, policy_id=policy_id):
             raise exceptions.PolicyUnknown
+        if not self.get_actions(user_id=user_id, policy_id=policy_id, perimeter_id=action_id):
+            raise exceptions.ActionUnknown
+        if not Managers.ModelManager.get_action_categories(user_id=user_id, category_id=category_id):
+            raise exceptions.MetaDataUnknown
+        if not self.get_action_data(user_id=user_id, policy_id=policy_id, data_id=data_id):
+            raise exceptions.DataUnknown
         return self.driver.add_action_assignment(policy_id=policy_id, action_id=action_id,
                                                  category_id=category_id, data_id=data_id)
 
@@ -271,11 +325,14 @@ class PolicyManager(Managers):
     @enforce("read", "rules")
     def get_rules(self, user_id, policy_id, meta_rule_id=None, rule_id=None):
         return self.driver.get_rules(policy_id=policy_id, meta_rule_id=meta_rule_id, rule_id=rule_id)
+        logger.info("delete_subject_data: {} {}".format(policy_id, data_id))
 
     @enforce(("read", "write"), "rules")
     def add_rule(self, user_id, policy_id, meta_rule_id, value):
         if not self.get_policies(user_id=user_id, policy_id=policy_id):
             raise exceptions.PolicyUnknown
+        if not self.ModelManager.get_meta_rules(user_id=user_id, meta_rule_id=meta_rule_id):
+            raise exceptions.MetaRuleUnknown
         return self.driver.add_rule(policy_id=policy_id, meta_rule_id=meta_rule_id, value=value)
 
     @enforce(("read", "write"), "rules")
